feat(ras): sync class
