+++ includes/class-newspack.php
@@ -81,8 +81,9 @@ final class Newspack {
 		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-activation-emails.php';
 		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-activation.php';
 		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-data.php';
-		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-esp-sync.php';
-		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-esp-sync-admin.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/sync/class-sync.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/sync/class-esp-sync.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/sync/class-esp-sync-admin.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-utils.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-data-events.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-webhooks.php';
