+++ includes/data-events/class-memberships.php
@@ -235,7 +235,7 @@ final class Memberships {
 	 */
 	public static function woocommerce_checkout_order_processed( $order_id ) {
 		$order = \wc_get_order( $order_id );
-		if ( ! \Newspack\WooCommerce_Connection::should_sync_order( $order ) ) {
+		if ( ! Reader_Activation\Sync::should_sync_order( $order ) ) {
 			return;
 		}
 		$data = self::get_order_data( $order_id, $order );
