+++ includes/reader-activation/sync/class-esp-sync.php
@@ -9,14 +9,13 @@ namespace Newspack\Reader_Activation;
 
 use Newspack\Reader_Activation;
 use Newspack\Logger;
-use Newspack\WooCommerce_Connection;
 
 defined( 'ABSPATH' ) || exit;
 
 /**
  * ESP Sync Class.
  */
-abstract class ESP_Sync {
+abstract class ESP_Sync extends Sync {
 
 	/**
 	 * Context of the sync.
@@ -177,7 +176,7 @@ abstract class ESP_Sync {
 			$customer->save();
 		}
 
-		$contact = $is_order ? WooCommerce_Connection::get_contact_from_order( $order ) : WooCommerce_Connection::get_contact_from_customer( $customer );
+		$contact = $is_order ? static::get_contact_from_order( $order ) : static::get_contact_from_customer( $customer );
 		if ( $registration_site ) {
 			$contact['metadata']['network_registration_site'] = $registration_site;
 		}
